==> Artificial Corner - Proyecto desde 0/1 Extrayendo Grupos AH 2022.ipynb <==
# 1 Demostración de Importación de tablas

import pandas as pd
from string import ascii_uppercase as alphabet #se importa el alfabeto en mayúsculas
import pickle

todas_tablas = pd.read_html('https://en.wikipedia.org/wiki/2022_FIFA_World_Cup')

# A -> H
# 9 -> 7 * 8 + 9 = 65
todas_tablas[9]
todas_tablas[16]
todas_tablas[23]

# Demostración de función zip en un bucle for
for letter, i in zip(alphabet, range (9, 65, 7)):
    print(letter, i)

# 2 Extrayendo y limpiando grupos 2022

def clear_group(df):
    # Cambiamos el nombre de la columna 'Teamvte' a 'Team'
    # Se utiliza df.columns[1] para ubicar la columna 'Teamvte' ya que no en todas las tablas se denomina igual
    df.rename(columns={df.columns[1]: 'Team'}, inplace=True)
    
    for col in df.columns[2:]:
        df[col].values[:] = 0
    
    return df
    

dict_grupos = {} 
for letter, i in zip(alphabet, range (9, 65, 7)): #zip permite trabajar con dos parámetros a la vez, en este caso el alfabeto
    df = todas_tablas[i]
    df = clear_group(df)
    
    #La columna 'Qualification' es igual en todas las tablas, por lo que no es necesario ubicarla por su posición
    df.pop('Qualification')
    #Creación del diccionario con los nombre de los grupos asignados de forma dinámica
    dict_grupos[f'Group {letter}'] = df

dict_grupos['Group H']

# Exportación del diccionario mediante la librería "pickle"
with open ('DictGroups2022', 'wb') as output: #with open (name_of_export, write_mode)
    pickle.dump(dict_grupos, output)
