==> Artificial Corner - Proyecto desde 0/2 Extrayendo Resultados de los Partidos por Año.ipynb <==
from bs4 import BeautifulSoup
import requests

import pandas as pd

# Todos los años en los que hubo mundial
years = [1930, 1934, 1938, 1950, 1954, 1958, 1962, 1970, 1974, 1978, 1982, 1986, 1990, 1994, 1998, 2002, 2006, 2010, 2014, 2018]

def get_matches(year):
    # Indicando el sitio web, el año del mundial y el modo de lectura lxml para el web scraping
    web = f'https://en.wikipedia.org/wiki/{year}_FIFA_World_Cup'
    response = requests.get(web)
    content = response.text
    soup = BeautifulSoup(content, 'lxml')
    
    # Extrayendo datos de cada partido por año
    matches = soup.find_all('div', class_='footballbox')
    home = []
    away = []
    score = []
    for match in matches:
        home.append(match.find('th', class_='fhome').get_text())
        away.append(match.find('th', class_='faway').get_text())
        score.append(match.find('th', class_='fscore').get_text())
    
    dict_football = {'home': home, 'away': away, 'score': score}
    df_football = pd.DataFrame(dict_football)
    df_football['year'] = year
    
    # Retornamos el dataframe
    return df_football

# Llamamos a la función, creamos una lista con cada dataframes y los concatenamos en uno sólo

fifa = [get_matches(year) for year in years] # Lista de comprensión, se resume el proceso a una sola línea
df_fifa = pd.concat(fifa, ignore_index=True) # Concatenamos los dataframes, es igual a decir pd.concat([df1, df2...])

df_fifa.to_csv('fifa_worldcup_historical_data.csv', index=False) # index=False no muestra los índices de cada registro

# Extraemos el fixture del año 2022 (sin los resultados)
df_fixture = get_matches(2022)
df_fixture.to_csv('fifa_worldcup_fixture.csv', index=False)
